% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EMD.R
\docType{package}
\name{emdomics-package}
\alias{emdomics-package}
\title{Earth Mover's Distance algorithm for differential analysis of genomics data.}
\description{
\code{\link{calculate_emd}} will usually be the only function needed.
}

